+++ src/emit.ts
